move test section header body mismatch test to a separate file
